backend/requirements.txt
Flask
Flask-SocketIO
opencv-python
mediapipe
Pillow
